Refactor SnapshotSensors and SnapshotSensorsApi to improve structure and readability; add SensorProducer and MetricsPrinter implementations for enhanced sensor data handling.
